constructors & to string

+++ src/Product.java
@@ -1,21 +1,21 @@
 public class Product extends Function{
 
-    public double valueAt(int x) {
+    public double valueAt(double x) {
 
     }
     public String derivative() {
 
     }
-    public double bisectionMethod(int a, int b, double epsilon) {
+    public double bisectionMethod(double a, double b, double epsilon) {
 
     }
-    public double bisectionMethod(int a, int b) {
+    public double bisectionMethod(double a, double b) {
 
     }
-    public double newtonRaphsonMethod(int a, double epsilon) {
+    public double newtonRaphsonMethod(double a, double epsilon) {
 
     }
-    public double newtonRaphsonMethod(int a) {
+    public double newtonRaphsonMethod(double a) {
 
     }
     public double taylorPolynomial(int n) {
